+++ packages/designer/src/managers/built-in/deps.ts
@@ -58,21 +58,21 @@ export const builtInDeps: Dependencie[] = [
     assetsLibrary: 'UniH5Material',
     platform: 'uniapp'
   },
-  {
-    package: 'uni-ui',
-    version: 'latest',
-    library: 'UniUI',
-    urls: [
-      '@vtj/materials/deps/uni-ui/style.css',
-      '@vtj/materials/deps/uni-ui/index.umd.js'
-    ],
-    required: true,
-    official: true,
-    enabled: true,
-    assetsUrl: '@vtj/materials/assets/uni-ui/index.umd.js',
-    assetsLibrary: 'UniUIMaterial',
-    platform: 'uniapp'
-  },
+  // {
+  //   package: 'uni-ui',
+  //   version: 'latest',
+  //   library: 'UniUI',
+  //   urls: [
+  //     '@vtj/materials/deps/uni-ui/style.css',
+  //     '@vtj/materials/deps/uni-ui/index.umd.js'
+  //   ],
+  //   required: true,
+  //   official: true,
+  //   enabled: true,
+  //   assetsUrl: '@vtj/materials/assets/uni-ui/index.umd.js',
+  //   assetsLibrary: 'UniUIMaterial',
+  //   platform: 'uniapp'
+  // },
   {
     package: '@vtj/utils',
     version: 'latest',